Moved around some tests
